remove nerdamer
This might be a temporary change. For the moment, I'm reluctant
for us to have too many dependencies in this project.

Removing nerdamer now lets us get the rest of the priority code
into the repo without further delays.

We can however discuss the benefits of nerdamer, and consider adding
it back.